Stubbing of work and minor improvements
