+++ src/lib.rs
@@ -1,12 +1,12 @@
 use gdnative::prelude::*;
 
 //mod logger;
-mod extensions;
-mod player;
+//mod extensions.rs;
+mod player_controls;
 
 fn init(handle: InitHandle) {
 	//handle.add_class::<logger::Logger>();
-	handle.add_class::<player::Player>();
+	handle.add_class::<player_controls::Player>();
 }
 
 godot_init!(init);
